Add settlement leg creation proof.

+++ dart-common/src/lib.rs
@@ -8,10 +8,10 @@ pub const MAX_AMOUNT: u64 = (1 << AMOUNT_BITS) - 1;
 
 pub const MAX_ASSET_ID: u32 = u32::MAX;
 
-pub const ACCOUNT_TREE_L: u32 = 512;
+pub const ACCOUNT_TREE_L: usize = 512;
 pub const ACCOUNT_TREE_HEIGHT: u8 = 4;
 pub const ACCOUNT_TREE_GENS: usize = 1 << 12;
 
-pub const ASSET_TREE_L: u32 = 64;
+pub const ASSET_TREE_L: usize = 64;
 pub const ASSET_TREE_HEIGHT: u8 = 4;
 pub const ASSET_TREE_GENS: usize = 1 << 13;
